Remove c2rust boilerplate: casts, extern duplication, struct centralization
- Remove redundant `size_of::<T>() as size_t` casts (87 occurrences)
- Remove `offset_from() as c_long as T` chain casts (12 occurrences)
- Remove `as c_int as c_long` chain casts (5 occurrences)
- Remove redundant `N as size_t` literal casts (175 occurrences)
- Remove redundant `N as u32/u64/u8` literal casts (~540 occurrences)
- Remove redundant `N as c_ulong as u16` chain casts (10 occurrences)
- Remove `as c_int` from const definitions (160 occurrences)
- Shorten `::core::mem::size_of` to `core::mem::size_of` (144 occurrences)
- Shorten `::core::ptr::null_mut` to `core::ptr::null_mut` (175 occurrences)
- Remove null_mut type parameters where inferable (147 occurrences)
- Replace `null_mut::<c_void>() as uintptr_t` with `0` (28 occurrences)
- Centralize `lzma_allocator` struct in types.rs (removed from 52 files)
- Centralize extern fn declarations in types.rs (lzma_alloc, lzma_free,
  memcpy, memset, memmove, memcmp, memchr, strlen)
- Remove unused `#![allow]` entries from lib.rs, auto-fix unused imports
- Update liblzma-rs-sys to use types::lzma_allocator

+++ liblzma-rs/src/lz/lz_encoder_mf.rs
@@ -1,5 +1,5 @@
 use crate::types::*;
-use core::ffi::{c_int, c_long, c_uint};
+use core::ffi::{c_int, c_uint};
 extern "C" {
     static lzma_crc32_table: [[u32; 256]; 8];
 }
@@ -128,7 +128,7 @@ unsafe extern "C" fn move_pos(mf: *mut lzma_mf) {
         (*mf).cyclic_pos = 0;
     }
     (*mf).read_pos = (*mf).read_pos.wrapping_add(1);
-    if ((*mf).read_pos.wrapping_add((*mf).offset) == 4294967295) as c_int as c_long != 0 {
+    if ((*mf).read_pos.wrapping_add((*mf).offset) == 4294967295) as c_int != 0 {
         normalize(mf);
     }
 }
@@ -226,7 +226,7 @@ pub unsafe extern "C" fn lzma_mf_hc3_find(mf: *mut lzma_mf, matches: *mut lzma_m
         matches.offset(matches_count as isize),
         len_best,
     )
-    .offset_from(matches) as c_long as u32;
+    .offset_from(matches) as u32;
     move_pos(mf);
     return matches_count;
 }
@@ -332,7 +332,7 @@ pub unsafe extern "C" fn lzma_mf_hc4_find(mf: *mut lzma_mf, matches: *mut lzma_m
         matches.offset(matches_count as isize),
         len_best,
     )
-    .offset_from(matches) as c_long as u32;
+    .offset_from(matches) as u32;
     move_pos(mf);
     return matches_count;
 }
@@ -509,7 +509,7 @@ pub unsafe extern "C" fn lzma_mf_bt2_find(mf: *mut lzma_mf, matches: *mut lzma_m
         matches.offset(matches_count as isize),
         1,
     )
-    .offset_from(matches) as c_long as u32;
+    .offset_from(matches) as u32;
     move_pos(mf);
     return matches_count;
 }
@@ -610,7 +610,7 @@ pub unsafe extern "C" fn lzma_mf_bt3_find(mf: *mut lzma_mf, matches: *mut lzma_m
         matches.offset(matches_count as isize),
         len_best,
     )
-    .offset_from(matches) as c_long as u32;
+    .offset_from(matches) as u32;
     move_pos(mf);
     return matches_count;
 }
@@ -747,7 +747,7 @@ pub unsafe extern "C" fn lzma_mf_bt4_find(mf: *mut lzma_mf, matches: *mut lzma_m
         matches.offset(matches_count as isize),
         len_best,
     )
-    .offset_from(matches) as c_long as u32;
+    .offset_from(matches) as u32;
     move_pos(mf);
     return matches_count;
 }
